requirements.txt
numpy
scikit-image
torch
torchvision
pillow
opencv-python
paddlepaddle
paddlehub
gradio
